+++ src/main/java/com/example/domain/mapping/StudyPost.java
@@ -1,6 +1,8 @@
-package com.example.domain.study;
+package com.example.domain.mapping;
 
 import com.example.domain.common.BaseEntity;
+import com.example.domain.enums.Theme;
+import com.example.domain.study.Study;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -9,6 +11,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
+import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -25,16 +28,27 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class StudyPostComment extends BaseEntity {
+public class StudyPost extends BaseEntity {
 
     @Id @GeneratedValue
     private long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    private StudyPost studyPost;
+    //private Member member;
 
-    // @Column(nullable = false)
-    // private Member member;
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Study study;
+
+    @Column(nullable = false)
+    private boolean isAnnouncement;
+
+    @Column(nullable = false)
+    private LocalDateTime announcedAt;
+
+    @Column(nullable = false)
+    private Theme theme;
+
+    @Column(nullable = false)
+    private String title;
 
     @Column(nullable = false)
     private String content;
@@ -42,11 +56,16 @@ public class StudyPostComment extends BaseEntity {
     @Column(nullable = false)
     private Integer likeCount;
 
-    @OneToMany(mappedBy = "studyPostComment", cascade = CascadeType.ALL)
-    private List<StudyLikedComment> likedComments;
+    @Column(nullable = false)
+    private Integer commentCount;
 
+    @OneToMany(mappedBy = "studyPost", cascade = CascadeType.ALL)
+    private List<StudyPostImage> images;
 
+    @OneToMany(mappedBy = "studyPost", cascade = CascadeType.ALL)
+    private List<StudyPostComment> comments;
 
-
+    @OneToMany(mappedBy = "studyPost", cascade = CascadeType.ALL)
+    private List<StudyLikedPost> likedPosts;
 
 }
